Move system_info.py into RimSort.utils
