+++ RimSort/controllers/app_controller.py
@@ -8,7 +8,7 @@ from PySide6.QtWidgets import QApplication
 
 from RimSort.utils.app_info import AppInfo
 from util.constants import DEFAULT_USER_RULES
-from util.system_info import SystemInfo
+from RimSort.utils.system_info import SystemInfo
 from view.main_window import MainWindow
